feat: add resource option to use pre existing secret
The bootstrap resource always writes the credentials used for
bootstrapping also to Secrets in flux namespace. This means it's not
possible to provision flux with read only acces to git repo this way.

Introduce new resource option `disable_secret_creation`, which, when set
to true causes the resource not to attempt to write the secret to Kubernetes
Secrets.

This secret must exist and it's not validated. As a caveat, if the
secret doesn't exist, a working one won't be created, but this is due to
how flux2/bootstrap works.

Signed-off-by: Lassi Pölönen <lassi.polonen@iki.fi>