Disable upgrade test
Revise the upgrade test setup to use a previous version of the provider
for the first bootstrap and use the current version to upgrade.
The flux2/v2 package used by the provider may not support the previous
API versions due to which installing a previous version of flux may not
succeed.

Signed-off-by: Sunny <darkowlzz@protonmail.com>