Remove version from the example used in test
Using the version of the provider that's not published from the example
configuration in e2e test fails because `terraform init` queries the
upstream terraform registry for the same version and fails when the
upstream returns an error, as it's unpublished. Remove the version field
from the example configuration to avoid this failure. The
`dev_overrides` in .terraformrc will make sure that the e2e test uses
the local provider instead of the downloaded old provider.

Signed-off-by: Sunny <darkowlzz@protonmail.com>